backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.0
faker==22.0.0

# API clients
tweepy==4.14.0
praw==7.7.1
requests==2.31.0
httpx==0.26.0
gql[all]==3.5.0

# AI/ML
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.3.1
spacy==3.7.2
bertopic==0.16.0
scikit-learn==1.4.0

# Blockchain
web3==6.15.0
solana==0.30.2
solders==0.18.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
aioredis==2.0.1
